+++ src/tools/__init__.py
@@ -9,7 +9,6 @@ Usage:
     from src.tools.expression_style import ExpressionStyleManager
     from src.tools.interrupt_question_tool import InterruptQuestionTool
     from src.tools.schedule_intent_tool import ScheduleIntentTool
-    from src.tools.tooltip_utils import ToolTip, create_treeview_tooltip
     from src.tools.settings_migration import SettingsMigration
 """
 
@@ -19,6 +18,5 @@ __all__ = [
     'expression_style',
     'interrupt_question_tool',
     'schedule_intent_tool',
-    'tooltip_utils',
     'settings_migration',
 ]